+++ src/components/SideBar/SideBarComponents/searchComponet.js
@@ -7,12 +7,12 @@ import {
 import IconButton from '@material-ui/core/IconButton';
 import Autocomplete, { createFilterOptions } from '@material-ui/lab/Autocomplete';
 import CloseIcon from '@material-ui/icons/Close';
-import ClearFilters from './clearFilters';
 import { getAllIds, localSearch, setSideBarToLoading } from '../../../pages/dashboardTab/store/dashboardReducer';
 import {
   search,
   defaultSearch,
 } from '../../../bento/dashboardData';
+import styles from './styles/searchComponentStyles';
 
 function getSearchResultColor(index, currentSection) {
   return index % 2 ? search[currentSection] ? search[currentSection].checkBoxColorsTwo ? search[currentSection].checkBoxColorsTwo : '' : defaultSearch.checkBoxColorsTwo
@@ -24,34 +24,41 @@ function getSearchResultCrossColor(currentSection) {
   return crossColor;
 }
 
-function localSearchCOmponent({ classes }) {
+const LocalSearchComponent = ({ classes }, ref) => {
   const [open, setOpen] = React.useState(false);
+  const [value, setValue] = React.useState([]);
   const [options, setOptions] = React.useState([]);
   const loading = open && options.length === 0;
-  const [value, setValue] = React.useState([]);
+
   const isSidebarLoading = useSelector((state) => (
     state.dashboardTab
   && state.dashboardTab.setSideBarLoading
       ? state.dashboardTab.setSideBarLoading : false));
+
   const tabDataLoading = useSelector((state) => (state.dashboardTab
     && state.dashboardTab.isDashboardTableLoading
     ? state.dashboardTab.isDashboardTableLoading
     : false));
   // redux use actions
 
+  React.useImperativeHandle(ref, () => ({
+    clear() {
+      setValue([]);
+      localSearch([]);
+      return null;
+    },
+  }));
+
   React.useEffect(() => {
     let active = true;
-
     if (!loading) {
       return undefined;
     }
-
     (async () => {
       const response = await getAllIds();
       const keys = ['fileIds', 'fileNames', 'sampleIds', 'subjectIds'];
       const mapOption = keys.map((key) => response[key].map((id) => ({ type: key, title: id })));
       const option = mapOption.reduce((acc = [], iterator) => [...acc, ...iterator]);
-
       if (active) {
         setOptions(option);
       }
@@ -75,29 +82,14 @@ function localSearchCOmponent({ classes }) {
     setValue(newValueUnique);
     localSearch(newValueUnique);
   }
+
   const onDelete = (title) => () => {
     const newValue = value.filter((v) => v.title !== title);
     onChange(newValue);
   };
 
-  function resetFilter() {
-    setValue([]);
-    localSearch([]);
-    return null;
-  }
   return (
     <>
-      {/* This is a temp solution for clear all need to find betetr solution
-       why clear filter on click not working */}
-      <div className={classes.clearFiltersBorder}>
-        <div>
-          <ClearFilters
-            disable={value.length === 0}
-            onClick={() => resetFilter()}
-            resetText="Clear all search selections"
-          />
-        </div>
-      </div>
       <div>
         <div className={classes.searchBoxRoot}>
           <Autocomplete
@@ -139,8 +131,8 @@ function localSearchCOmponent({ classes }) {
         </div>
         <Divider
           style={{
-            backgroundColor: '#B0CFE1',
             height: '1px',
+            backgroundColor: '#B0CFE1',
           }}
         />
         <div>
@@ -187,56 +179,8 @@ function localSearchCOmponent({ classes }) {
       </Backdrop>
     </>
   );
-}
+};
 
-const styles = () => ({
+const localSearchComponent = React.forwardRef(LocalSearchComponent);
 
-  backdrop: {
-    // position: 'absolute',
-    zIndex: 99999,
-    background: 'rgba(0, 0, 0, 0.1)',
-  },
-  clearFiltersBorder: {
-    borderTop: '1px solid black',
-    borderBottom: '1px solid #B0CFE1',
-  },
-  inputRoot: {
-    borderRadius: 10,
-    // maxHeight: 10,
-    marginTop: '5px',
-    marginBottom: '5px',
-    '& .MuiOutlinedInput-notchedOutline': {
-      borderColor: '#0F5B9C',
-    },
-    '&:hover .MuiOutlinedInput-notchedOutline': {
-      borderColor: '#0F5B9C',
-    },
-    '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-      borderColor: '#0F5B9C',
-    },
-    '& .MuiInputBase-input': {
-      height: '7px',
-    },
-  },
-  closeRoot: {
-    marginLeft: '3px',
-    height: 18,
-  },
-  listItemGutters: {
-    padding: '2px 0px 2px 0px',
-  },
-  searchResultDetailText: {
-    marginTop: '1.5px',
-    color: '#000000',
-    lineHeight: '120%',
-    fontFamily: 'Nunito',
-    fontSize: '14px',
-  },
-  searchBoxRoot: {
-    marginLeft: 'Auto',
-    marginRight: 'Auto',
-    width: '90%',
-  },
-});
-
-export default withStyles(styles)(localSearchCOmponent);
+export default withStyles(styles)(localSearchComponent);
