+++ src/pages/dashboardTab/dashboard.js
@@ -11,6 +11,7 @@ import ActiveFiltersQuery from '../../components/ActiveFiltersQuery/ActiveFilter
 import { widgetsData, displayActiveFiltersQuery } from '../../bento/dashboardData';
 import Tab from './components/tabController';
 import colors from '../../utils/colors';
+import styles from './dashboardStyles';
 
 const displaywidgets = widgetsData.filter((widget) => widget.show === true).slice(0, 6);
 
@@ -18,11 +19,8 @@ const Dashboard = ({
   classes, data, theme,
 }) => {
   const [collapse, setCollapse] = React.useState(true);
-  const [activeTabId, setActiveTabId] = React.useState(null);
   const themeChanger = useTheme();
-  const handleChange = () => {
-    setCollapse((prev) => !prev);
-  };
+  const handleChange = () => setCollapse((prev) => !prev);
 
   return (
     <>
@@ -31,11 +29,11 @@ const Dashboard = ({
         <div>
           <div className={classes.content}>
             <div className={classes.sideBar}>
-              <SideBar setActiveTabId={setActiveTabId} />
+              <SideBar />
             </div>
             <div className={classes.rightContent}>
               <div className={classes.widgetsContainer}>
-                {displayActiveFiltersQuery && activeTabId === '1' ? <ActiveFiltersQuery /> : ''}
+                {displayActiveFiltersQuery ? <ActiveFiltersQuery /> : ''}
                 <div className={classes.widgetsCollapse}>
                   <div className={classes.floatLeft} />
                   <div className={classes.floatRight}>
@@ -44,7 +42,7 @@ const Dashboard = ({
                         <Button className={classes.customButton} onClick={handleChange}>
                           {collapse ? 'COLLAPSE VIEW' : 'OPEN VIEW' }
                         </Button>
-)}
+                      )}
                     />
                     <Switch
                       classes={{
@@ -141,113 +139,5 @@ const Dashboard = ({
     </>
   );
 };
-
-const styles = (theme) => ({
-  dashboardContainer: {
-    backgroundColor: '#FFFFFF',
-    // boxShadow: 'inset 0 0 87px 7px rgba(27,28,28,0.15)',
-  },
-  dashboardDivider: {
-    height: 16,
-    marginTop: '32px',
-    backgroundColor: '#E2E7EC',
-  },
-  dashboardDividerTop: {
-    height: 16,
-    backgroundColor: theme.palette.widgetBackground.main,
-  },
-  rightContent: {
-    maxWidth: 'calc(100% - 250px)',
-    position: 'relative',
-  },
-  content: {
-    // padding: theme.spacing.unit * 3,
-    display: 'flex',
-    maxWidth: '1800px',
-    margin: 'auto',
-  },
-  widgetsContainer: {
-    background: theme.palette.widgetBackground.main,
-  },
-  contentShift: {
-    width: `calc(100vw - ${theme.custom.drawerWidth})`,
-    marginLeft: theme.custom.drawerWidth,
-    transition: theme.transitions.create(['margin', 'width'], {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  card: {
-    minHeight: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  paper: {
-    textAlign: 'center',
-  },
-  fakeToolbar: {
-    ...theme.mixins.toolbar,
-  },
-  sunburst: {
-    textAlign: 'center',
-  },
-  widgetInner: {
-    marginTop: '-8px',
-    borderBottom: '6px solid #E2E7EC',
-  },
-  widgetsCollapse: {
-    background: theme.palette.widgetBackground.main,
-  },
-  floatRight: {
-    float: 'right',
-    marginRight: '80px',
-  },
-  floatLeft: {
-    float: 'left',
-  },
-  customSwitch: {
-    marginTop: '-6px',
-  },
-  customButton: {
-    borderRadius: '0 0 18px 18px',
-    minHeight: '20px',
-    fontSize: 8,
-    color: '#ffffff',
-    textTransform: 'none',
-    backgroundColor: '#566672',
-    marginRight: '4px',
-    fontFamily: theme.custom.fontFamilySans,
-    marginTop: '-4px',
-    '&:hover': {
-      backgroundColor: '#566672',
-    },
-  },
-  backgroundWidgets: {
-    background: theme.palette.widgetBackground.main,
-  },
-  sideBar: {
-    width: '250px',
-    overflowX: 'hidden',
-    backgroundColor: 'transparent',
-    borderRight: 'thin solid #B1B1B1',
-    borderLeft: 'thin solid #B1B1B1',
-    overflow: 'auto',
-    zIndex: '99',
-  },
-  statsBar: {
-    position: 'fixed',
-  },
-  switchBase: {
-    color: theme.palette.widgetBackground.contrastText,
-    '&$checked': {
-      color: theme.palette.widgetBackground.contrastSwicthColor,
-    },
-    '&$checked + $track': {
-      backgroundColor: theme.palette.widgetBackground.contrastText,
-    },
-  },
-  checked: {},
-  track: {},
-});
 
 export default withStyles(styles, { withTheme: true })(Dashboard);
